# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: Yoobe V3 API - Sistema de Orçamentos e Replicação
  description: |
    API para gerenciamento de orçamentos e replicação de produtos base.
    
    ## Fluxo de Orçamento → Aprovação → Replicação
    
    1. **Gestor** cria orçamento com produtos desejados
    2. **Admin Global** aprova ou rejeita o orçamento
    3. **Gestor** pode replicar produtos apenas após aprovação
    4. **Gestor** gerencia produtos replicados (ativar/inativar, editar)
    
    ## Autenticação
    
    Todas as rotas requerem autenticação via Bearer Token.
    - Role `admin`: Acesso total ao sistema
    - Role `manager`: Acesso limitado às funcionalidades de gestor
    - Role `user`: Acesso apenas a funcionalidades básicas
    
  version: 2.1.0
  contact:
    name: Yoobe Team
    email: support@yoobe.co
servers:
  - url: http://localhost:3000
    description: Servidor de Desenvolvimento
  - url: https://api.yoobe.co
    description: Servidor de Produção
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BaseProduct:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        base_price:
          type: number
          format: float
        base_points_cost:
          type: integer
        image_url:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        category_id:
          type: string
          format: uuid
          nullable: true
    CompanyProduct:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        points_cost:
          type: integer
        stock_quantity:
          type: integer
        is_active:
          type: boolean
        company_id:
          type: string
          format: uuid
        base_product_id:
          type: string
          format: uuid
        budget_id:
          type: string
          format: uuid
        approved_at:
          type: string
          format: date-time
        approved_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        error:
          type: string
          description: Mensagem de erro
        code:
          type: string
          description: Código de erro (opcional)
        details:
          type: object
          description: Detalhes adicionais do erro
paths:
...
  /api/gestor/base-products:
    get:
      summary: Listar produtos base disponíveis
      description: Retorna produtos base que podem ser replicados
      tags: [Gestor - Produtos Base]
      security:
        - BearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Termo de busca
        - name: category
          in: query
          schema:
            type: string
            format: uuid
          description: Filtrar por categoria
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Lista de produtos base
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/BaseProduct'
                        - type: object
                          properties:
                            is_replicated:
                              type: boolean
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      totalPages:
                        type: integer

    post:
      summary: Replicar produto base
      description: Replica um produto base para a loja do gestor (requer orçamento aprovado)
      tags: [Gestor - Produtos Base]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - base_product_id
              properties:
                base_product_id:
                  type: string
                  format: uuid
                custom_price:
                  type: number
                  format: float
                  minimum: 0
                custom_points_cost:
                  type: integer
                  minimum: 0
                custom_stock_quantity:
                  type: integer
                  minimum: 0
                custom_description:
                  type: string
                  maxLength: 1000
      responses:
        '201':
          description: Produto replicado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  company_product:
                    $ref: '#/components/schemas/CompanyProduct'
                  message:
                    type: string
        '403':
          description: Orçamento não aprovado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ========================================
  # GESTOR - PRODUTOS REPLICADOS
  # ========================================
...